# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type IqraChatRequestMessagesItem = {
  role: string;
  content: string;
};

export type IqraChatRequestMode = typeof IqraChatRequestMode[keyof typeof IqraChatRequestMode];


export const IqraChatRequestMode = {
  teacher: 'teacher',
  student: 'student',
} as const;

export type IqraChatRequestLanguage = typeof IqraChatRequestLanguage[keyof typeof IqraChatRequestLanguage];


export const IqraChatRequestLanguage = {
  ar: 'ar',
  en: 'en',
} as const;

export interface IqraChatRequest {
  messages: IqraChatRequestMessagesItem[];
  /** Knowledge-base context (lesson text) to ground the answer */
  context?: string;
  mode?: IqraChatRequestMode;
  language?: IqraChatRequestLanguage;
}

export type GenerateRequestLanguage = typeof GenerateRequestLanguage[keyof typeof GenerateRequestLanguage];


export const GenerateRequestLanguage = {
  arabic: 'arabic',
  english: 'english',
} as const;

export type GenerateRequestTeachingStyle = typeof GenerateRequestTeachingStyle[keyof typeof GenerateRequestTeachingStyle];


export const GenerateRequestTeachingStyle = {
  direct: 'direct',
  inquiry: 'inquiry',
  collaborative: 'collaborative',
} as const;

export type GenerateRequestDifficulty = typeof GenerateRequestDifficulty[keyof typeof GenerateRequestDifficulty];


export const GenerateRequestDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
  mixed: 'mixed',
} as const;

export interface GenerateRequest {
  grade: string;
  subject: string;
  topic: string;
  duration?: number;
  language?: GenerateRequestLanguage;
  additionalContext?: string;
  teachingStyle?: GenerateRequestTeachingStyle;
  objectives?: string;
  difficulty?: GenerateRequestDifficulty;
  numQuestions?: number;
  questionTypes?: string[];
  totalMarks?: number;
}

export interface LessonPlanOutput {
  title: string;
  grade: string;
  subject: string;
  duration: number;
  objectives: string[];
  materials: string[];
  introduction: string;
  mainActivity: string;
  guidedPractice: string;
  independentPractice: string;
  closure: string;
  assessment: string;
  differentiation: string;
  homework: string;
}

export interface WorksheetQuestion {
  text: string;
  options?: string[];
  answer?: string;
  points: number;
}

export interface WorksheetSection {
  type: string;
  title: string;
  questions: WorksheetQuestion[];
}

export interface WorksheetAnswerKeyItem {
  num: number;
  answer: string;
}

export interface WorksheetOutput {
  title: string;
  instructions: string;
  sections: WorksheetSection[];
  answerKey: WorksheetAnswerKeyItem[];
}

export interface QuizQuestion {
  id: string;
  type: string;
  text: string;
  options?: string[];
  correctAnswer: string;
  points: number;
  explanation: string;
}

export interface QuizOutput {
  title: string;
  duration: number;
  totalPoints: number;
  questions: QuizQuestion[];
}