# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /generate/lesson-plan:
    post:
      operationId: generateLessonPlan
      tags: [generate]
      summary: Generate an AI lesson plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "200":
          description: Generated lesson plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonPlanOutput"
        "500":
          description: Generation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /generate/worksheet:
    post:
      operationId: generateWorksheet
      tags: [generate]
      summary: Generate an AI worksheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "200":
          description: Generated worksheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorksheetOutput"
        "500":
          description: Generation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /generate/quiz:
    post:
      operationId: generateQuiz
      tags: [generate]
      summary: Generate an AI quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "200":
          description: Generated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizOutput"
        "500":
          description: Generation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
...
components:
  schemas:
...
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    GenerateRequest:
      type: object
      properties:
        grade:
          type: string
        subject:
          type: string
        topic:
          type: string
        duration:
          type: integer
        language:
          type: string
          enum: [arabic, english]
        additionalContext:
          type: string
        teachingStyle:
          type: string
          enum: [direct, inquiry, collaborative]
        objectives:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, mixed]
        numQuestions:
          type: integer
        questionTypes:
          type: array
          items:
            type: string
        totalMarks:
          type: integer
      required:
        - grade
        - subject
        - topic
    LessonPlanOutput:
      type: object
      properties:
        title:
          type: string
        grade:
          type: string
        subject:
          type: string
        duration:
          type: integer
        objectives:
          type: array
          items:
            type: string
        materials:
          type: array
          items:
            type: string
        introduction:
          type: string
        mainActivity:
          type: string
        guidedPractice:
          type: string
        independentPractice:
          type: string
        closure:
          type: string
        assessment:
          type: string
        differentiation:
          type: string
        homework:
          type: string
      required:
        - title
        - grade
        - subject
        - duration
        - objectives
        - materials
        - introduction
        - mainActivity
        - guidedPractice
        - independentPractice
        - closure
        - assessment
        - differentiation
        - homework
    WorksheetSection:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/WorksheetQuestion"
      required:
        - type
        - title
        - questions
    WorksheetQuestion:
      type: object
      properties:
        text:
          type: string
        options:
          type: array
          items:
            type: string
        answer:
          type: string
        points:
          type: integer
      required:
        - text
        - points
    WorksheetAnswerKeyItem:
      type: object
      properties:
        num:
          type: integer
        answer:
          type: string
      required:
        - num
        - answer
    WorksheetOutput:
      type: object
      properties:
        title:
          type: string
        instructions:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/WorksheetSection"
        answerKey:
          type: array
          items:
            $ref: "#/components/schemas/WorksheetAnswerKeyItem"
      required:
        - title
        - instructions
        - sections
        - answerKey
    QuizQuestion:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        points:
          type: integer
        explanation:
          type: string
      required:
        - id
        - type
        - text
        - correctAnswer
        - points
        - explanation
    QuizOutput:
      type: object
      properties:
        title:
          type: string
        duration:
          type: integer
        totalPoints:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
      required:
        - title
        - duration
        - totalPoints
        - questions