# lib/api-zod/src/generated/types/generateRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateRequestDifficulty } from './generateRequestDifficulty';
import type { GenerateRequestLanguage } from './generateRequestLanguage';
import type { GenerateRequestTeachingStyle } from './generateRequestTeachingStyle';

export interface GenerateRequest {
  grade: string;
  subject: string;
  topic: string;
  duration?: number;
  language?: GenerateRequestLanguage;
  additionalContext?: string;
  teachingStyle?: GenerateRequestTeachingStyle;
  objectives?: string;
  difficulty?: GenerateRequestDifficulty;
  numQuestions?: number;
  questionTypes?: string[];
  totalMarks?: number;
}