# lib/api-zod/src/generated/types/quizQuestion.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuizQuestion {
  id: string;
  type: string;
  text: string;
  options?: string[];
  correctAnswer: string;
  points: number;
  explanation: string;
}