# lib/api-zod/src/generated/types/worksheetAnswerKeyItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorksheetAnswerKeyItem {
  num: number;
  answer: string;
}