# lib/api-zod/src/generated/types/worksheetOutput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorksheetAnswerKeyItem } from './worksheetAnswerKeyItem';
import type { WorksheetSection } from './worksheetSection';

export interface WorksheetOutput {
  title: string;
  instructions: string;
  sections: WorksheetSection[];
  answerKey: WorksheetAnswerKeyItem[];
}