# lib/api-zod/src/generated/types/worksheetQuestion.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorksheetQuestion {
  text: string;
  options?: string[];
  answer?: string;
  points: number;
}