# lib/api-zod/src/generated/types/worksheetSection.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorksheetQuestion } from './worksheetQuestion';

export interface WorksheetSection {
  type: string;
  title: string;
  questions: WorksheetQuestion[];
}